requirements.txt
pyboy
pillow
openai
python-dotenv
